+++ packages/react-native-reanimated/Common/cpp/reanimated/RuntimeDecorators/RNRuntimeDecorator.cpp
@@ -1,6 +1,6 @@
 #include <reanimated/RuntimeDecorators/RNRuntimeDecorator.h>
 #include <reanimated/Tools/ReanimatedVersion.h>
-#include <worklets/Tools/JSIUtils.h>
+#include <worklets/Tools/WorkletsJSIUtils.h>
 
 namespace reanimated {
